pyproject.toml
[tool.poetry]
name = "google-pubsub-emulator-python-async2"
version = "0.1.0"
description = ""
authors = ["Daniele Esposti <daniele.esposti@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
google-api-core = "^1.26.1"
google-auth = "^1.28.0"
google-cloud-pubsub = "^2.4.0"
grpcio = "^1.36.1"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
isort = "^5.7.0"
black = "^20.8b1"
pdbpp = "^0.10.2"
ipython = "^7.21.0"
safety = "^1.10.3"
flake8 = "^3.9.0"
autoflake = "^1.4"
grpc-stubs = "^1.24.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
